Split encrypted attribute concerns into 2 modules
**Why**: Makes attribute declarations explicit, shorter files.
