+++ app/models/concerns/user_encrypted_attribute_overrides.rb
@@ -1,10 +1,6 @@
 module UserEncryptedAttributeOverrides
   extend ActiveSupport::Concern
 
-  attr_accessor :attribute_user_access_key
-
-  ENCRYPTED_ATTRIBUTES = [:phone, :otp_secret_key].freeze
-
   # override some Devise methods to support our use of encrypted_email
 
   class_methods do
@@ -50,88 +46,9 @@ module UserEncryptedAttributeOverrides
     generate_confirmation_token
   end
 
-  def email
-    get_encrypted_attribute(name: :email, default: '')
-  end
-
+  # Override usual setter method in order to also set fingerprint
   def email=(email)
     set_encrypted_attribute(name: :email, value: email, default: '')
     self.email_fingerprint = email.present? ? encrypted_attributes[:email].fingerprint : ''
-  end
-
-  def stale_encrypted_email?
-    return false unless email.present?
-    encrypted_attributes[:email].stale?
-  end
-
-  ENCRYPTED_ATTRIBUTES.each do |attribute|
-    class_eval <<-METHODS, __FILE__, __LINE__ + 1
-      def #{attribute}
-        get_encrypted_attribute(name: :"#{attribute}", default: nil)
-      end
-    METHODS
-  end
-
-  ENCRYPTED_ATTRIBUTES.each do |attribute|
-    class_eval <<-METHODS, __FILE__, __LINE__ + 1
-      def #{attribute}=(attribute)
-        set_encrypted_attribute(name: :"#{attribute}", value: attribute, default: nil)
-      end
-
-      def stale_encrypted_#{attribute}?
-        return false unless self.public_send(:"#{attribute}")
-        encrypted_attributes[:"#{attribute}"].stale?
-      end
-    METHODS
-  end
-
-  private
-
-  def encrypted_attributes
-    @_encrypted_attributes ||= {}
-  end
-
-  def get_encrypted_attribute(name:, default:)
-    getter = encrypted_attribute_name(name)
-    encrypted_string = self[getter]
-    return default unless encrypted_string.present?
-    encrypted_attribute = encrypted_attributes[name]
-    return encrypted_attribute.decrypted if encrypted_attribute.present?
-    build_encrypted_attribute(name, encrypted_string).decrypted
-  end
-
-  def build_encrypted_attribute(name, encrypted_string)
-    encrypted_attribute = EncryptedAttribute.new(
-      encrypted_string,
-      cost: attribute_cost,
-      user_access_key: attribute_user_access_key
-    )
-    self.attribute_user_access_key ||= encrypted_attribute.user_access_key
-    encrypted_attributes[name] = encrypted_attribute
-  end
-
-  def set_encrypted_attribute(name:, value:, default:)
-    setter = encrypted_attribute_name(name)
-    new_value = default
-    if value.present?
-      new_value = build_encrypted_attribute_from_plain(name, value).encrypted
-    end
-    self[setter] = new_value
-  end
-
-  def build_encrypted_attribute_from_plain(name, plain_value)
-    self.attribute_user_access_key ||= new_attribute_user_access_key
-    encrypted_attributes[name] = EncryptedAttribute.new_from_decrypted(
-      plain_value.downcase.strip,
-      attribute_user_access_key
-    )
-  end
-
-  def encrypted_attribute_name(name)
-    "encrypted_#{name}".to_sym
-  end
-
-  def new_attribute_user_access_key
-    EncryptedAttribute.new_user_access_key(cost: attribute_cost)
   end
 end
